Add files via upload
This is a website to search the weather condition based on the user’s geographic location and give recommendation for clothing and food using the DarkSky Weather API.
Our website will have interactive features for user to enter the location and country. These information will be added by clicking button or pressing enter. Once the user enter their location and click the enter button, our website will return the temperature, condition, humidity, recommendation for clothes, and recommendation for food.